+++ render/src/drawquad.rs
@@ -25,7 +25,7 @@ impl Clone for DrawQuad{
 
 impl DrawQuad {
     pub fn new(cx: &mut Cx) -> Self {
-        Self::with_shader(cx, live_shader!(cx, self::bg_shader), 0)
+        Self::with_shader(cx, live_shader!(cx, self::shader), 0)
     }
     
     pub fn with_shader(_cx: &mut Cx, shader: Shader, slots: usize) -> Self {
@@ -37,6 +37,49 @@ impl DrawQuad {
             rect: Rect::default(),
             z: 0.0
         }
+    }
+    
+    pub fn style(cx: &mut Cx){
+        live_body!(cx, r#"
+            
+            self::shader: Shader {
+                
+                use crate::shader_std::prelude::*;
+                
+                default_geometry: crate::shader_std::quad_2d;
+                geometry geom: vec2;
+                
+                varying pos: vec2;
+                
+                instance x: float;
+                instance y: float;
+                instance w: float;
+                instance h: float;
+                instance z: float;
+                
+                //let dpi_dilate: float<Uniform>;
+                fn scroll() -> vec2 {
+                    return draw_scroll.xy;
+                }
+                
+                fn vertex() -> vec4 {
+                    let scr = scroll();
+                    
+                    let clipped: vec2 = clamp(
+                        geom * vec2(w, h) + vec2(x, y) - scr,
+                        draw_clip.xy,
+                        draw_clip.zw
+                    );
+                    pos = (clipped + scr - vec2(x, y)) / vec2(w, h);
+                    // only pass the clipped position forward
+                    return camera_projection * (camera_view * (view_transform * vec4(clipped.x, clipped.y, z + draw_zbias, 1.)));
+                }
+                
+                fn pixel() -> vec4 {
+                    return #0f0;
+                }
+            }
+        "#);        
     }
     
     pub fn begin_quad(&mut self, cx: &mut Cx, layout: Layout) {
@@ -115,4 +158,3 @@ impl DrawQuad {
         }
     }
 }
-
