+++ render/src/drawtext.rs
@@ -1,202 +1,265 @@
-
 use crate::cx::*;
 
-#[derive(Copy, Clone)]
-pub enum Wrapping {
-    Char,
-    Word,
-    Line,
-    None,
-    Ellipsis(f32)
-}
-
-/*
-#[derive(Clone, Copy)]
-pub struct TextStyle {
-    pub font: Font,
-    pub font_size: f32,
-    pub brightness: f32,
-    pub curve: f32,
-    pub line_spacing: f32,
-    pub top_drop: f32,
-    pub height_factor: f32,
-}
-
-impl Default for TextStyle {
-    fn default() -> Self {
-        TextStyle {
-            font: Font::default(),
-            font_size: 8.0,
-            brightness: 1.0,
-            curve: 0.6,
-            line_spacing: 1.4,
-            top_drop: 1.1,
-            height_factor: 1.3,
-        }
-    }
-}*/
-
-#[derive(Clone)]
-pub struct Text {
-    pub text_style: TextStyle,
+#[repr(C, packed)]
+pub struct DrawText {
     pub shader: Shader,
-    pub color: Color,
-    pub z: f32,
+    pub area: Area,
+    pub lock: Option<LockedInstances>,
+    pub slots: usize,
+    pub chunk: Vec<char>,
+    pub text_style: TextStyle,
     pub wrapping: Wrapping,
     pub font_scale: f32,
+    pub z: f32,
+    
+    pub font_tx1: f32,
+    pub font_ty1: f32,
+    pub font_tx2: f32,
+    pub font_ty2: f32,
+    pub color: Color,
+    pub rect: Rect,
+    pub z_bias: f32,
+    pub base_x: f32,
+    pub base_y: f32,
+    pub font_size: f32,
+    pub char_offset: f32,
+    pub marker: f32,
 }
 
-impl Text {
-    pub fn new(cx: &mut Cx) -> Self {
+
+impl Clone for DrawText {
+    fn clone(&self) -> Self {
         Self {
-            text_style: TextStyle {
-                font: Font{font_id:0},
+            shader: unsafe {self.shader.clone()},
+            area: Area ::Empty,
+            lock: None,
+            slots: self.slots,
+            chunk: Vec::new(),
+
+            text_style: self.text_style,
+            wrapping: self.wrapping,
+            font_scale: self.font_scale,
+            z: self.z,
+
+            font_tx1: 0.0,
+            font_ty1: 0.0,
+            font_tx2: 0.0,
+            font_ty2: 0.0,
+            color: self.color,
+            rect: Rect::default(),
+            z_bias: 0.0,
+            base_x: 0.0,
+            base_y: 0.0,
+            font_size: 0.0,
+            char_offset: 0.0,
+            marker: 0.0,
+        }
+    }
+}
+
+impl DrawText {
+    
+    pub fn new(cx: &mut Cx) -> Self {
+        Self::with_shader(cx, live_shader!(cx, self::shader), 0)
+    }
+    
+    pub fn with_shader(cx: &mut Cx, shader: Shader, slots: usize) -> Self {
+        Self {
+            shader: shader,
+            area: Area::Empty,
+            lock: None,
+            slots: slots + 18,
+            chunk: Vec::new(),
+            
+            text_style: live_text_style!(cx, self::text_style_unscaled),/*TextStyle {
+                font: Font {font_id: 0},
                 font_size: 8.0,
                 brightness: 1.0,
                 curve: 0.6,
                 line_spacing: 1.4,
                 top_drop: 1.1,
                 height_factor: 1.3,
-            },
-            shader: live_shader!(cx, self::shader),
-            z: 0.0,
+            },*/
             wrapping: Wrapping::Word,
-            color: Color::parse_name("white").unwrap(),
             font_scale: 1.0,
+            z: 0.0,
+            
+            font_tx1: 0.0,
+            font_ty1: 0.0,
+            font_tx2: 0.0,
+            font_ty2: 0.0,
+            color: Color::parse_name("white").unwrap(),
+            rect: Rect::default(),
+            z_bias: 0.0,
+            base_x: 0.0,
+            base_y: 0.0,
+            font_size: 0.0,
+            char_offset: 0.0,
+            marker: 0.0,
         }
     }
     
     pub fn style(cx: &mut Cx) {
         
-        live_body!(cx, r#"self::shader: Shader {
-
-            use crate::shader_std::prelude::*;
-            
-            default_geometry: crate::shader_std::quad_2d;
-            
-            geometry geom: vec2;
-
-            texture texturez: texture2D;
-            
-            instance font_tc: vec4;
-            instance color: vec4;
-            instance x: float;
-            instance y: float;
-            instance w: float;
-            instance h: float;
-            instance z: float;
-            instance base_x: float;
-            instance base_y: float;
-            instance font_size: float;
-            instance char_offset: float;
-            instance marker: float;
-            
-            varying tex_coord1: vec2;
-            varying tex_coord2: vec2;
-            varying tex_coord3: vec2;
-            varying clipped: vec2;
-            //let rect: vec4<Varying>;
-            
-            uniform brightness: float;
-            uniform curve: float;
-            
-            fn get_color() -> vec4 {
-                return color;
+        live_body!(cx, r#"
+            self::text_style_unscaled: TextStyle {
+                font: "resources/Ubuntu-R.ttf",
+                font_size: 8.0,
+                brightness: 1.0,
+                curve: 0.6,
+                line_spacing: 1.4,
+                top_drop: 1.2,
+                height_factor: 1.3,
             }
             
-            fn pixel() -> vec4 {
+            self::shader: Shader {
                 
-                let dx = dFdx(vec2(tex_coord1.x * 2048.0, 0.)).x;
-                let dp = 1.0 / 2048.0;
-
-                // basic hardcoded mipmapping so it stops 'swimming' in VR
-                // mipmaps are stored in red/green/blue channel
-                let s = 1.0;
-                if dx > 7.0 {
-                    s = 0.7;
-                }
-                else if dx > 2.75 {
-                    s = (
-                        sample2d(texturez, tex_coord3.xy + vec2(0., 0.)).z
-                            + sample2d(texturez, tex_coord3.xy + vec2(dp, 0.)).z
-                            + sample2d(texturez, tex_coord3.xy + vec2(0., dp)).z
-                            + sample2d(texturez, tex_coord3.xy + vec2(dp, dp)).z
-                    ) * 0.25;
-                }
-                else if dx > 1.75 {
-                    s = sample2d(texturez, tex_coord3.xy).z;
-                }
-                else if dx > 1.3 {
-                    s = sample2d(texturez, tex_coord2.xy).y;
-                }
-                else {
-                    s = sample2d(texturez, tex_coord1.xy).x;
+                use crate::shader_std::prelude::*;
+                
+                default_geometry: crate::shader_std::quad_2d;
+                
+                geometry geom: vec2;
+                
+                texture texturez: texture2D;
+                
+                instance font_tc: vec4;
+                instance color: vec4;
+                instance x: float;
+                instance y: float;
+                instance w: float;
+                instance h: float;
+                instance z: float;
+                instance base_x: float;
+                instance base_y: float;
+                instance font_size: float;
+                instance char_offset: float;
+                instance marker: float;
+                
+                varying tex_coord1: vec2;
+                varying tex_coord2: vec2;
+                varying tex_coord3: vec2;
+                varying clipped: vec2;
+                //let rect: vec4<Varying>;
+                
+                uniform brightness: float;
+                uniform curve: float;
+                
+                fn get_color() -> vec4 {
+                    return color;
                 }
                 
-                s = pow(s, curve);
-                let col = get_color();//color!(white);//get_color();
-                return vec4(s * col.rgb * brightness * col.a, s * col.a); 
+                fn pixel() -> vec4 {
+                    let dx = dFdx(vec2(tex_coord1.x * 2048.0, 0.)).x;
+                    let dp = 1.0 / 2048.0;
+                    
+                    // basic hardcoded mipmapping so it stops 'swimming' in VR
+                    // mipmaps are stored in red/green/blue channel
+                    let s = 1.0;
+                    if dx > 7.0 {
+                        s = 0.7;
+                    }
+                    else if dx > 2.75 {
+                        s = (
+                            sample2d(texturez, tex_coord3.xy + vec2(0., 0.)).z
+                                + sample2d(texturez, tex_coord3.xy + vec2(dp, 0.)).z
+                                + sample2d(texturez, tex_coord3.xy + vec2(0., dp)).z
+                                + sample2d(texturez, tex_coord3.xy + vec2(dp, dp)).z
+                        ) * 0.25;
+                    }
+                    else if dx > 1.75 {
+                        s = sample2d(texturez, tex_coord3.xy).z;
+                    }
+                    else if dx > 1.3 {
+                        s = sample2d(texturez, tex_coord2.xy).y;
+                    }
+                    else {
+                        s = sample2d(texturez, tex_coord1.xy).x;
+                    }
+                    
+                    s = pow(s, curve);
+                    let col = get_color(); //color!(white);//get_color();
+                    return vec4(s * col.rgb * brightness * col.a, s * col.a);
+                }
+                
+                fn vertex() -> vec4 {
+                    let min_pos = vec2(x, y);
+                    let max_pos = vec2(x + w, y - h);
+                    
+                    clipped = clamp(
+                        mix(min_pos, max_pos, geom) - draw_scroll.xy,
+                        draw_clip.xy,
+                        draw_clip.zw
+                    );
+                    
+                    let normalized: vec2 = (clipped - min_pos + draw_scroll.xy) / vec2(w, -h);
+                    //rect = vec4(min_pos.x, min_pos.y, max_pos.x, max_pos.y) - draw_scroll.xyxy;
+                    
+                    tex_coord1 = mix(
+                        font_tc.xy,
+                        font_tc.zw,
+                        normalized.xy
+                    );
+                    
+                    tex_coord2 = mix(
+                        font_tc.xy,
+                        font_tc.xy + (font_tc.zw - font_tc.xy) * 0.75,
+                        normalized.xy
+                    );
+                    
+                    tex_coord3 = mix(
+                        font_tc.xy,
+                        font_tc.xy + (font_tc.zw - font_tc.xy) * 0.6,
+                        normalized.xy
+                    );
+                    
+                    return camera_projection * (camera_view * (view_transform * vec4(clipped.x, clipped.y, z + draw_zbias, 1.)));
+                }
             }
-            
-            fn vertex() -> vec4 {
-                let min_pos = vec2(x, y);
-                let max_pos = vec2(x + w, y - h);
-                
-                clipped = clamp(
-                    mix(min_pos, max_pos, geom) - draw_scroll.xy,
-                    draw_clip.xy,
-                    draw_clip.zw
-                );
-                
-                let normalized: vec2 = (clipped - min_pos + draw_scroll.xy) / vec2(w, -h);
-                //rect = vec4(min_pos.x, min_pos.y, max_pos.x, max_pos.y) - draw_scroll.xyxy;
-                
-                tex_coord1 = mix(
-                    font_tc.xy,
-                    font_tc.zw,
-                    normalized.xy
-                );
-                
-                tex_coord2 = mix(
-                    font_tc.xy,
-                    font_tc.xy + (font_tc.zw - font_tc.xy) * 0.75,
-                    normalized.xy
-                );
-                
-                tex_coord3 = mix(
-                    font_tc.xy,
-                    font_tc.xy + (font_tc.zw - font_tc.xy) * 0.6,
-                    normalized.xy
-                );
-                
-                return camera_projection * (camera_view * (view_transform * vec4(clipped.x, clipped.y, z + draw_zbias, 1.)));
-            }
-        }"#);
+        "#);
     }
     
-    pub fn begin_text(&mut self, cx: &mut Cx) -> AlignedInstance {
-        //let font_id = self.font.font_id.unwrap();
-        let inst = cx.new_instance(self.shader, None, 0);
-        let aligned = cx.align_instance(inst);
-        let text_style = &self.text_style;
-        let brightness = text_style.brightness;
-        let curve = text_style.curve;
-        if aligned.inst.is_first_instance() {
-            aligned.inst.write_texture_2d_id(cx, live_item_id!(self::shader::texturez), cx.fonts_atlas.texture_id);
-            aligned.inst.write_uniform_float(cx, live_item_id!(self::shader::brightness), brightness);
-            aligned.inst.write_uniform_float(cx, live_item_id!(self::shader::curve), curve);
+    pub fn lock_aligned_text(&mut self, cx: &mut Cx) {
+        self.lock = Some(cx.lock_aligned_instances(self.shader, self.slots));
+    }
+    
+    pub fn lock_text(&mut self, cx: &mut Cx) {
+        self.lock = Some(cx.lock_instances(self.shader, self.slots));
+    }
+    
+    pub fn unlock_text(&mut self, cx: &mut Cx) {
+        unsafe {
+            if let Some(li) = self.lock.take() {
+                let offset = li.instance_area.instance_offset;
+                self.area = cx.unlock_instances(li);
+                if offset == 0 {
+                    self.area.write_texture_2d_id(cx, live_item_id!(self::shader::texturez), cx.fonts_atlas.texture_id);
+                    self.area.write_uniform_float(cx, live_item_id!(self::shader::brightness), self.text_style.brightness);
+                    self.area.write_uniform_float(cx, live_item_id!(self::shader::curve), self.text_style.curve);
+                }
+            }
         }
-        return aligned
     }
     
-    pub fn add_text<F>(&mut self, cx: &mut Cx, geom_x: f32, geom_y: f32, char_offset: usize, aligned: &mut AlignedInstance, chunk: &[char], mut char_callback: F)
+    pub fn add_text(&mut self, cx: &mut Cx, geom_x: f32, geom_y: f32, val:&str){
+        let mut chunk = Vec::new();
+        std::mem::swap(&mut chunk, unsafe{&mut self.chunk});
+        chunk.truncate(0);
+        for c in val.chars(){
+            chunk.push(c)
+        }
+        self.add_text_chunk(cx, geom_x, geom_y, 0, &chunk, | _, _, _, _ | {0.0});
+        std::mem::swap(&mut chunk, unsafe{&mut self.chunk});
+    }
+    
+    pub fn add_text_chunk<F>(&mut self, cx: &mut Cx, geom_x: f32, geom_y: f32, char_offset: usize, chunk: &[char], mut char_callback: F)
     where F: FnMut(char, usize, f32, f32) -> f32
     {
-        if geom_x.is_nan() || geom_y.is_nan(){
+        if geom_x.is_nan() || geom_y.is_nan() {
             return
         }
         
-        let text_style = &self.text_style;
+        let text_style = unsafe {&self.text_style};
+        
         let mut geom_x = geom_x;
         let mut char_offset = char_offset;
         let font_id = text_style.font.font_id;
@@ -215,20 +278,17 @@ impl Text {
         
         let atlas_page = &mut cxfont.atlas_pages[atlas_page_id];
         
-        let instances = {
-            let cxview = &mut cx.views[aligned.inst.view_id];
-            let draw_call = &mut cxview.draw_calls[aligned.inst.draw_call_id];
-            &mut draw_call.instances
-        };
+        let li = unsafe {if let Some(li) = &mut self.lock {li} else {return}};
         
         for wc in chunk {
+            
             let unicode = *wc as usize;
             let glyph_id = font.char_code_to_glyph_index_map[unicode];
             if glyph_id >= font.glyphs.len() {
                 println!("GLYPHID OUT OF BOUNDS {} {} len is {}", unicode, glyph_id, font.glyphs.len());
                 continue;
             }
-
+            
             let glyph = &font.glyphs[glyph_id];
             
             let advance = glyph.horizontal_metrics.advance_width * font_size_logical * self.font_scale;
@@ -283,49 +343,46 @@ impl Text {
             };
             
             // give the callback a chance to do things
-            let marker = char_callback(*wc, char_offset, geom_x, advance);
+            self.font_tx1 = tc.tx1;
+            self.font_ty1 = tc.ty1;
+            self.font_tx2 = tc.tx2;
+            self.font_ty2 = tc.ty2;
+            self.rect = Rect {
+                x: scaled_min_pos_x,
+                y: scaled_min_pos_y,
+                w: w * self.font_scale / dpi_factor,
+                h: h * self.font_scale / dpi_factor
+            };
+            self.z_bias = self.z + 0.00001 * min_pos_x;
+            self.base_x = geom_x;
+            self.base_y = geom_y;
+            self.font_size = text_style.font_size;
+            self.char_offset = char_offset as f32;
             
-            let data = [
-                tc.tx1,
-                tc.ty1,
-                tc.tx2,
-                tc.ty2,
-                self.color.r, // color
-                self.color.g,
-                self.color.b,
-                self.color.a,
-                scaled_min_pos_x,
-                scaled_min_pos_y,
-                w * self.font_scale / dpi_factor,
-                h * self.font_scale / dpi_factor,
-                self.z + 0.00001 * min_pos_x, //slight z-bias so we don't get z-fighting with neighbouring chars overlap a bit
-                geom_x,
-                geom_y,
-                text_style.font_size,
-                char_offset as f32, // char_offset
-                marker, // marker
-            ];
-            instances.extend_from_slice(&data);
+            // self.marker = marker;
+            self.marker = char_callback(*wc, char_offset, geom_x, advance);
+
+            li.instances.extend_from_slice(unsafe {
+                std::slice::from_raw_parts(&self.font_tx1 as *const _ as *const f32, self.slots)
+            });
             // !TODO make sure a derived shader adds 'empty' values here.
             
             geom_x += advance;
             char_offset += 1;
-            aligned.inst.instance_count += 1;
         }
     }
     
-    pub fn end_text(&mut self, cx: &mut Cx, aligned: &AlignedInstance) -> Area {
-        cx.update_aligned_instance_count(aligned);
-        aligned.inst.into()
-    }
-    
-    pub fn draw_text(&mut self, cx: &mut Cx, text: &str) -> Area {
-        let mut aligned = self.begin_text(cx);
+    pub fn draw_text(&mut self, cx: &mut Cx, text: &str) {
+        self.lock_aligned_text(cx);
         
         let mut chunk = Vec::new();
+        std::mem::swap(&mut chunk, unsafe{&mut self.chunk});
+        chunk.truncate(0);
+        
         let mut width = 0.0;
         let mut elipct = 0;
-        let text_style = &self.text_style;
+        
+        let text_style = unsafe {&self.text_style};
         let font_size = text_style.font_size;
         let line_spacing = text_style.line_spacing;
         let height_factor = text_style.height_factor;
@@ -393,7 +450,8 @@ impl Text {
                     margin: Margin::zero()
                 });
                 
-                self.add_text(cx, geom.x, geom.y, 0, &mut aligned, &chunk, | _, _, _, _ | {0.0});
+                self.add_text_chunk(cx, geom.x, geom.y, 0, &chunk, | _, _, _, _ | {0.0});
+                
                 width = 0.0;
                 chunk.truncate(0);
                 if newline {
@@ -401,11 +459,14 @@ impl Text {
                 }
             }
         }
-        self.end_text(cx, &aligned)
+        std::mem::swap(&mut chunk, unsafe{&mut self.chunk});
+        
+        self.unlock_text(cx);
     }
     
     // looks up text with the behavior of a text selection mouse cursor
-    pub fn find_closest_offset(&self, cx: &Cx, area: &Area, pos: Vec2) -> usize {
+    pub fn find_closest_offset(&self, cx: &Cx, pos: Vec2) -> usize {
+        let area = unsafe{&self.area};
         let scroll_pos = area.get_scroll_pos(cx);
         let spos = Vec2 {x: pos.x + scroll_pos.x, y: pos.y + scroll_pos.y};
         let x_o = area.get_instance_offset(cx, live_item_id!(self::shader::base_x), Ty::Float).unwrap();
@@ -414,7 +475,7 @@ impl Text {
         let font_size_o = area.get_instance_offset(cx, live_item_id!(self::shader::font_size), Ty::Float).unwrap();
         let char_offset_o = area.get_instance_offset(cx, live_item_id!(self::shader::char_offset), Ty::Float).unwrap();
         let read = area.get_read_ref(cx);
-        let text_style = &self.text_style;
+        let text_style = unsafe {&self.text_style};
         let line_spacing = text_style.line_spacing;
         let mut index = 0;
         if let Some(read) = read {
@@ -433,7 +494,7 @@ impl Text {
                             let prev_w = read.instances[read.offset + w_o + prev_index * read.slots];
                             if index < read.count - 1 && prev_x > spos.x + prev_w { // fix newline jump-back
                                 return read.instances[read.offset + char_offset_o + index * read.slots] as usize;
-                            } 
+                            }
                             return read.instances[read.offset + char_offset_o + prev_index * read.slots] as usize;
                         }
                         index += 1;
